Added a view to retrieve all twisks for a certain tag

+++ twisker/views.py
@@ -32,7 +32,11 @@ def twisks():
 @app.route('/twisks/<string:id>', methods=['GET', 'DELETE'])
 @json_query_view
 def get_twisk(id):
+    """Returns a single twisk, also accept DELETE"""
     twisk = Twisk.get_by_id(int(id))
+
+    if not twisk:
+        abort(404)
 
     if request.method == 'DELETE':
         twisk.delete()
@@ -40,3 +44,9 @@ def get_twisk(id):
         return twisk.content
 
 
+@app.route('/twisks/tagged/<string:tag>', methods=['GET'])
+@json_query_view
+def get_tag(tag):
+    """Returns all twisks tagged with a given tag"""
+    twisks = Twisk.gql("WHERE tags = :1 ORDER BY when DESC", tag)
+    return twisks
